Link component scaffold
