+++ packages/react/package.json
@@ -33,6 +33,7 @@
   ],
   "dependencies": {
     "@amsterdam/design-system-react-icons": "workspace:*",
+    "@utrecht/component-library-react": "1.0.0-alpha.296",
     "clsx": "1.2.1",
     "date-fns": "2.29.3"
   },
