+++ packages/react/src/Link/README.md
@@ -1,0 +1,3 @@
+# React Link Component
+
+[Link documentation](../../../css/src/link/README.md)
