+++ packages/react/src/Link/index.ts
@@ -1,0 +1,1 @@
+export { Link } from './Link'
