+++ packages/react/src/index.ts
@@ -4,7 +4,7 @@
  */
 
 export * from './Heading'
+export * from './Link'
 export { Icon } from './Icon'
-export { Link } from './Link'
 export { List } from './List'
 export { Paragraph } from './Paragraph'
